refactor trace store chunk read node helpers
